+++ texts/src/main/scala/org/opentorah/texts/tanach/VerseStore.scala
@@ -1,0 +1,5 @@
+package org.opentorah.texts.tanach
+
+import org.opentorah.store.Store
+
+class VerseStore(override val number: Int) extends Store.Numbered with Store.Terminal